Remove unused python code
